pyproject.toml
[tool.poetry]
name = "whisper-video-summarization"
version = "0.1.0"
description = "Whisper Video Summarization project with FastAPI, MLflow and Streamlit"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "whisper_video_summarization"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.124.4"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
requests = "^2.32.5"
pydantic = "^2.12.5"
python-multipart = "^0.0.21"
pyyaml = "^6.0.3"

[tool.poetry.group.inference.dependencies]
torch = "^2.6.0"
torchaudio = "^2.6.0"
torchvision = "^0.21.0"
openai-whisper = "*"
transformers = "^4.57.3"
pytorch-lightning = "^2.6.0"
sentence-transformers = "^5.2.0"
soundfile = "^0.13.1"
av = "^12.3.0"
sentencepiece = "*"

[tool.poetry.group.mlflow.dependencies]
mlflow = "^3.7.0"
sqlalchemy = "*"

[tool.poetry.group.streamlit.dependencies]
streamlit = "^1.52.1"
pandas = "^2.3.3"
numpy = "^1.26.4"
scikit-learn = "^1.8.0"
tqdm = "^4.67.1"

[tool.poetry.group.mlops.dependencies]
dvc = "*"
dvc-s3 = "*"
gitpython = "^3.1.45"
fsspec = "^2025.12.0"
torchmetrics = "^1.8.2"
rouge-score = "^0.1.2"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["whisper_video_summarization"]
